Bio::DB::Query::GenBank now correctly handles an explicitly uploaded list of GIs, locus numbers or accessions.
For the record, there is an inaccuracy in the eutils documentation.
The tools do not recognize a composite database named "sequences," but
require that "protein", "nucleotide" or "genome" be requested
explicitly.  Also epost is not very useful since it does not indicate
how many of the requested IDs were valid.

svn path=/bioperl-live/trunk/; revision=5126