see bp maillist for full desc
the script seq/unflatten_seq will now generate GFF3 - the unflattener
module is used to build the 'feature graph' connecting genes, transcripts,
exons and CDSs together. This means we can have GFF3 for anything in
genbank!

As far as I'm aware, the only other sensible output formats to use here
(ie formats that support feature graphs/containment hierarchies) are:
chado, chaos, and the write-only asciitree.

This feature graph is written out in the GFF3 using the ID and Parent
tags. To do this there is an extra intermediate step - the bioperl
FeatureHolderI hierarchy is traversed and ID/Parent tags are generated.

Here is a description of the changes I have made:

[unless you're a bioperl hacker you don't really need to read the rest of
this]

You can get the context of what I'm on about from this thread:
http://bioperl.org/pipermail/bioperl-l/2003-December/014150.html

Two new public methods:

  FeatureHolderI->set_ParentIDs_from_hierarchy

    sets both ID and ParentID from FeatureHolder hierarchy

  SeqFeatureI->generate_unique_persistent_id

    this is required by the above method

    Lincoln wanted this to be private, but I think it has
    to be called from outside

  FeatureHolderI->create_hierarchy_from_ParentIDs

    the inverse of set_ParentIDs_from_hierarchy

svn path=/bioperl-live/trunk/; revision=7714