added code which, if a file to read from doesn't exist and the name looks like a web url, attempts to download the file and use a local tempfile of it instead.  this allows:
Some::IO->new(-file=>'http://server.my/test.gff')

to pull down a gff file, cache it locally, and read from it.

it may be better to have a special -url option or something to keep this
code separated out, but this commit is just to get it out there.

svn path=/bioperl-live/trunk/; revision=8205