Since the bp_bulk_load_gff.pl has been replaced with a script that takes the place of both itself (mysql version) and bp_pg_bulk_load_gff.pl (postgresversion), we want to create a symbolic link from postgresversion to bp_bulk_load_gff.pl on install.  This will make the change transparent to people who have been using the postgres version.
On make, an external script maintenence/symlink_scripts.pl gets created with all the necessary path info.  In the postamble of Makefile.PL, a line was inserted to call the symlink_scripts.pl file.  Then on install, symlink_scripts.pl is run and creates the symbolic link.

The Perl symlink function was used to create the link, so that on systems where symlink doesn't work, it catches the error and prints a note tothe user.  That is untested though.

svn path=/bioperl-live/trunk/; revision=8837