added spaces to end of homology strings:
here's a weird parsing 'bug': In a blastn, one will have an hsp

100 atcg 103
    || |
400 atgg 397

No problem, the query string is 'atcg'
            the match string is '|| |'
            the hit   string is 'atgg'

But if we have

100 atcg 103
    |||
400 atcc 397

then        the query string is 'atcg'
	    the match string is '|||'
	    the hit   string is 'atcc'

That is, the match string is missing the space at the end. This 
may be a bug in the report generating program. 


svn path=/bioperl-live/trunk/; revision=15738