Update pod and make get_trace_data() return the current value.
The pod references the option -read_graph_data and the method
read_graph_data(), but neither are handled by the code; the code
uses get_trace_data.

The method get_trace_data() is used as an accessor in the code:
called without an argument to read the value - but the method
overwrites the current value with 0 if called without any arguments;
so calling get_trace_data() without arguments returns 0 always,
making it impossible to reach the read_trace_with_graph() call.