see Re: [Gmod-gbrowse] bp_genbank2gff3.pl bugs?
I've just taught Unflattener.pm about ribosomal slippages.

It's currently relatively strict, but passes on this example. I'm no expert on this phenomena - if you see one slippage do you expect to sometimes see more in the same translation? Is it always a frameshift of 1, or can it jump more?

also added a test file derived from (see OAZ3)
http://www.ncbi.nlm.nih.gov/nuccore/NT_004487.19?&from=3224086&to=3232447&report=genbank

and added a test to Unflattener.t


svn path=/bioperl-live/trunk/; revision=16459