Introspect feature objects and call apporpriate GFF output method
Depending on how the database was built, next_seq() may return
something that supports gff3_string() or someting that supports
gff_string(). Introspect on the feature object to figure out which of
these two methods are supported, and call the appropriate one:
gff3_string() if it exists, gff_string() otherwise.

If neither method exists on the feature object, thrown an exception
with a stack trace.