Added feature for deletion sites and bugfixes
'delete' method now adds a misc_feature with a note about
the length of the deletion site. The location type of this feature
is IN-BETWEEN.

Features of type IN-BETWEEN must have adjacent start/
end pos, so they are now deleted in the 'insert' method if they
co-localise with the insertion site. This happens when 'delete' is
followed by 'insert' or when using the 'ligate' shortcut method.
'insert' now also handles deleted features like 'delete', which
only applies to features with IN-BETWEEN locations.

Other fixes:

 - 'ligate' now skips the 'delete' step if 'left' and 'right' are
   adjacent because no deletion actually occurs.

 - '_coord_adjust_deletion': fixed test for splitting a feature.
   can not use 'contains' because that returns true if one or both
   coordinates of the feature and deletions co-localise but a split
   only makes sense when both ends overlap.

 - added and modified tests accordingly