PrimarySeq direct mode fix/improvement
A -direct option exists in PrimarySeq to skip any sequence validation.
However, the documentation lets one think that it was effective for
sequence strings passed to -seq whereas it actually only works when
passing a sequence string reference to -ref_to_seq.

I made it so that -direct is effective for sequences passed to -seq
or seq() as well, more in line with what the documentation states.