Constructor -ref_to_seq fix
The constructor -ref_to_seq to make a new Bio::PrimarySeq object based on a
string reference (instead of string) was not honored unless the -direct
constructor was also passed. This was not a documented behavior, and there is
no reason that -ref_to_seq has to be passed with -direct.

One can now use -ref_to_seq and -direct independently.