Cleaned messy file approach
Problem: when indexing an arbitrary set of file, there is no way to reconstruct a full filepath using path() and file()

Solution:
Save entire filepath instead of simply file basename in the object
Have file() return the basename of that (for legacy)
Have another function filepath() return the full filepath. This is much more useful than having the basename of the file anyway.