FastAlign.pl
I did a bioperl script which uses the heuristic method of tfasty.
The query string is in amino acids and the hit string is in nucleotides.
There are extra nucleotides at the end of the hit string (option -diff and by
default = 10), that allow to verify if the intron start with common rules
(5'-GTGCGA-... for group II intron and after an exonic T for group I intron).

I'm glad to help the bioperl community!