Add tests for spliced_seq remote sequence handling
Add tests to check that the length of spliced_seq is the same when it
resorts to padding with Ns after failing to locate the remote source
sequence for some exons. Also checks that spliced_seq throws the
appropriate warning. The test throws a warning itself if it cannot
complete the tests e.g. due to lack of Genbank access.