Add a personal script to extract a DNA sequence from a FASTA file using a BLAST output file. Expects at least two arguments, the BLAST file and the FASTA file.
This script is especially useful when trying to extract sequences with variance (hence the BLAST search beforehand) from FASTA files.

For example, say that you are trying to extract a given gene and 2000 base pairs 5' to it from 20 different genomes.  All you have is the gene sequence from one of those genomes, however.  By doing a BLAST search between each of the genomes and the gene you have and then using this script, you can extract the sequences that you are interested in.

The script also has options to extract a specified 3' or 5' sequence from the FASTA file, as well as an e-value cut off.

Output is the sequence in FASTA format.