Fixed issue #126: Parsing DT lines from embl files
From the issue #126: in embl.pm, lines 305 and 311, the two regex that parse the DT lines are always false do to the use of the modifier x without escaping white-space in regex. Thus 'creation_release', 'update_release' and 'update_version' are always missing. I proposed to remove the x modifiers.

This fix revealed an other issue while writing Bio::Seq into an embl file. The call of the thee above values is wrong and the output display the ref type of the variables instead of the values. To fix it, I simply add a call to the method ->value(), to get the value.