Avoid "Unescaped left brace in regex is deprecated" warning.
In Perl 5.24.0, when running t/Tools/SiRNA.t, we were getting the warning
above on two consecutive lines.  The underlying reason for the warnings was
that pattern in question never really was a regex quantifier.

This patch silences those warnings and
prevents the module from failing on future versions of perl.

The test suite was never exercising the code in question.  Now it will
exercise the code in question, but that will mean different numbers of pairs
to be expected in t/Tools/SiRNA.t.