Consistent shebang line to '+#!/usr/bin/perl'
During installation, the shebang line is automatically replaced with
the perl used for installation (assuming that the distribution was
done with bioperl's distzilla plugin bundle).  This makes sense
because that's the only perl we know for sure will have bioperl
installed.

However, MakeMaker does not replace the shebang line if it's using
/usr/bin/env perl.  So let's be consistent.