Remove entire Bio::PopGen namespace and related modules, scripts, and test
These are a modular unit so they are now a separate repo named
Bio-PopGen (not yet released).  The Bio::Tree::AlleleNode has also
been moved since it is related to Bio::PopGen, dependent on it, and
not used anywhere else.

Also moved the scripts bp_composite_LD and bp_heterogeneity_test.