Bio::Perl: move its own distribution named Bio-Procedural
This module is meant as a wrapper for the rest of bioperl.  It provides
a procedural interface for those that can't program in object oriented
programming.

Because it wraps multiple other modules, some of which we want to make
on their own distribution, we can't keep it in bioperl core without
mkaing it dependent on the modules we want to move out.