maint: simplify README to cover simple install instructions.
Remove instructions on how to install perl on different OS, there's
plenty of pages online on how to do it, nothing of this is BioPerl
specific.  Also link to the CPAN page about how to install perl
modules which applies to BioPerl modules.

Make reference to how BioPerl project is actually multiple
distributions and to install with cpanm by naming the module
required.

Add new HACKING.md, merge of README and INSTALL files, for users
interested in development.