t/LocalDB/Index/Blast.t: create test tempfiles within a tempdir (issue #281)
There's other tests that create files with the same name on the current
working directory with the same name.  If running the test suite with
jobs in parallel this causes failing tests, because multiple tests will
be writing and reading to the same file at the same time.

So use File::Temp to avoid clashes.

Specific case here is that Bio::Index::Blast->new will not creat a new
file, it will use an existing one if it's there. In this case, it was
reading an Index file for anoter format which had a different version
number.

Take the chance to rewrite the test to avoid code duplication.