maint: 644 permissions for almost all files
All files permissions set to 644.  Exception are scripts in maintenance
and travis_scripts (all set to 755) since they are meant to be run after
clone.

Even stuff in scripts/ is set to 644 because it's only meant to be run
after the installation, and never from the git repository (the release
and install steps both modify the scripts).