Bio::LiveSeq::* move whole namespace to its own distribution.
Everything moved to its own distribution named Bio-LiveSeq.  The
modules in this namespace are dependent on Bio::Variation.  With
this is a separate distribution, Bio::Variation can also be moved
without circular dependencies, and Bio::Variation brings in a lot
of extra dependencies.

This also adjust test in SeqUtils that uses a LiveSeq.