Remove Bio::SeqIO support for abi, alf, ctf, exp, pln, and ztr.
This trace files were wrappers to Bio::SeqIO::staden::read which is
not even indexed in CPAN (it's in the bioperl-ext repository).  So
remove this modules and move into a separate distribution (actually,
move them into the bioperl-ext repo).

Note that alf.t makes use of a test data file that no longer exists.
Looking into the history of the repo, seems that it never did.