Remove manipulation of @INC and use of lib - require install for use.
Some scripts manipulated @INC so they could be ran without installation.
Undo this, we expect users to have it installed.  The authoring and
installation step of BioPerl modifies the code and it may be required
to work properly.  Savy enough users can manually add the path when
calling the script if they want.

Do not catch failing to load a module to give a different error message.
The standard perl message about missing module in @INC should be enough.