Changes: more detailed complete list of modules/dependencies added/removed
Simply list all modules and dependencies that have been removed and added
for this new release.  The list of modules goes more in line with what the
users should focus on, individual modules and not the distribution.  Also,
we no longer say to what distribution a module went because that module
might change later.  This is simpler, less verbose, and more complete.