Bio::Root::RootI: deprecate deprecated() method (issue #294)
This method is not worth it.  If something is deprecated, then it is
deprecated then, not in a future version.  And then eventually gets
removed.  Either the code is there and runs with a warning about
deprecation, or the code is not there in which case it dies.

This removes some methods that were supposed to throw() starting
with 1.7.5 (which will be the next release).