Bio::Tools::CodonTable: include and parse the gc.prt file (#391)
To make it simpler to keep things up to date, just include the gc.prt
(the NCBI data file of codon tables in ASN.1 print format) below
__DATA__ and parse it.  The file is very small and parsing it is very
quick.

We can't read DATA in a BEGIN block so this also changes the BEGIN
into a INIT block.  I'm not sure why the code there needs to be inside
them though.