t/SeqTools/CodonTable.t: test for removed codon tables (#386)
Codon tables are sometimes removed, e.g., codon table id=7.
Previously, instead of being handled as invalid tables, these would
return the "''" (because they were parsed from a `qw()`.  Since
a73febe9 the tables are parsed straight from the gc.prt, the tables
array gets a real empty string, adn are handled as invalid.

In PR #386 a test was added that the "''" was returned.  This is no
longer needed.  This commit tests that when a table is empty (id=7)
it's properly identified as invalid to avoid a regression.