Bio::Tools::CodonTable::is_start_codon: check in case of ambiguous codons (#266)
In the case of ambiguous codons this method should only return true if
all possible codons are start codons (same as is_ter_codon).  This was
the behaviour in BioPerl 1.2.3 but since at least BioPerl 1.6.924 it
returns true if at least one codon is a start codon.

This change fixes that regression and makes behaviour consistent with
is_ter_codon.