dist.ini: ignore Bio::DB reqs only when coming from bin/bp_fetch
bin/bp_fetch requires a bunch of Bio::DB modules from other BioPerl
distributions which we want to suggest only (because it makes
installation more difficult).  Instead of always ignoring them, this
commit changes it so it only ignores them when AutoPrereqs finds those
reqs on bin/bp_fetch.  This is to prevent future issues of missed reqs
if we ever start requiring them on other modules where "suggesting" is
not appropriate.